Be 101 errors

+++ backend/models/users.models.js
@@ -33,7 +33,7 @@ exports.loginUser = async (username, password) => {
     const user = await User.findOne({username});
 
     if (!user) {
-        return Promise.reject(errors.errMsg_invalidItem('username'));
+        return Promise.reject(errors.errMsg_idNotFound);
     }
 
     const validPassword = await bcrypt.compare(password, user.password);
@@ -89,11 +89,11 @@ exports.deleteUser = async (userId, password) => {
         return Promise.reject(errors.errMsg_idNotFound);
     }
 
-    // const validPassword = await bcrypt.compare(password, user.password);
+    const validPassword = await bcrypt.compare(password, user.password);
 
-    // if (!validPassword){
-    //     return Promise.reject(errors.errMsg_invalidItem('password'));
-    // }
+    if (!validPassword){
+        return Promise.reject(errors.errMsg_invalidItem('password'));
+    }
 
     const res = await User.deleteOne({_id: userId});
